backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai==1.54.0
mss==9.0.1
Pillow==10.4.0
aiofiles==23.2.1
python-multipart==0.0.6
requests==2.31.0